rename and split classes into files
