+++ src/window_background/arena-log-watcher.js
@@ -48,7 +48,7 @@ import {
   setData,
   updateLoading
 } from "./background-util";
-import { playerDb } from "../shared/local-database";
+import { playerDb } from "../shared/LocalDatabase";
 import { ARENA_MODE_MATCH, ARENA_MODE_DRAFT } from "../shared/constants";
 import update_deck from "./updateDeck";
 import globals from "./globals";
