+++ src/window_background/background.js
@@ -4,7 +4,8 @@ import path from "path";
 import fs from "fs";
 import sha1 from "js-sha1";
 import * as httpApi from "./http-api";
-import { appDb, playerDb, rememberDefaults } from "../shared/local-database";
+import { appDb, playerDb } from "../shared/LocalDatabase";
+import { rememberDefaults } from "../shared/databaseUtil";
 import playerData from "../shared/player-data";
 import { getReadableFormat } from "../shared/util";
 import { HIDDEN_PW, MAIN_DECKS } from "../shared/constants";
