+++ src/window_main/settings.js
@@ -25,7 +25,7 @@ import {
   SETTINGS_LOGIN
 } from "../shared/constants";
 import db from "../shared/database";
-import { playerDefaults } from "../shared/local-database";
+import { playerDefaults } from "../shared/databaseUtil";
 import pd from "../shared/player-data";
 import {
   createDiv,
